🎉 init : 검색기능 추가, js 간단 리펙토링
